refactor: merge develop

+++ app/src/main/java/com/rocket/cosmic_detox/presentation/view/fragment/mypage/MyPageViewModel.kt
@@ -85,9 +85,9 @@ class MyPageViewModel @Inject constructor(
     fun setAppUsageLimit(allowedApp: AllowedApp, hour: String, minute: String) {
         viewModelScope.launch {
             //val limitedTime = (hour.toInt() * 60 + minute.toInt()) * 60 * 1000L // 시간과 분을 밀리초로 변환
-            val limitedTime = (hour.toInt() * 60 + minute.toInt()) * 60 // 시간과 분을 초로 변환
+            val limitedTime = (hour.toInt() * 60 + minute.toInt()) * 60L // 시간과 분을 초로 변환
 
-            repository.updateAppUsageLimit(allowedApp.copy(limitedTime = limitedTime.toInt()))
+            repository.updateAppUsageLimit(allowedApp.copy(limitedTime = limitedTime))
                 .onSuccess {
                     // TODO: 나중에 UiState로 변경해보기
                     _updateResult.value = true
